# Application/backend/swagger.yaml
openapi: 3.0.0
info:
  title: Quest - Fraud Detection API
  description: |
    A comprehensive fraud detection and risk monitoring system for financial transactions.
    The API provides endpoints for authentication, dashboard analytics, transaction management,
    network visualization, and report generation.
  version: 1.0.0
  contact:
    name: API Support
    email: support@quest-fraud.local
  license:
    name: MIT
servers:
  - url: http://localhost:5000/api
    description: Local development server
  - url: https://api.quest-fraud.com/api
    description: Production server
...
paths:
...
  /dashboard/stats:
    get:
      tags:
        - Dashboard
      summary: Get dashboard statistics
      description: Returns overall statistics for the fraud detection dashboard
      operationId: getDashboardStats
      responses:
        '200':
          description: Dashboard statistics retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  totalAccounts:
                    type: integer
                    example: 1250
                  totalTransactions:
                    type: integer
                    example: 45820
                  totalVolume:
                    type: number
                    format: double
                    example: 5234850.50
                  highRiskTransactions:
                    type: integer
                    example: 342
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
        details:
          type: string
        requestId:
          type: string
        timestamp:
          type: string
          format: date-time
...